requirements.txt
google-generativeai
pysrt
cryptography
pytest
